+++ certbot-apache/certbot_apache/_internal/override_fedora.py
@@ -1,10 +1,13 @@
 """ Distribution specific override class for Fedora 29+ """
-from certbot import errors
-from certbot import util
+from typing import Any
+
 from certbot_apache._internal import apache_util
 from certbot_apache._internal import configurator
 from certbot_apache._internal import parser
 from certbot_apache._internal.configurator import OsOptions
+
+from certbot import errors
+from certbot import util
 
 
 class FedoraConfigurator(configurator.ApacheConfigurator):
@@ -68,7 +71,7 @@ class FedoraConfigurator(configurator.ApacheConfigurator):
 
 class FedoraParser(parser.ApacheParser):
     """Fedora 29+ specific ApacheParser override class"""
-    def __init__(self, *args, **kwargs) -> None:
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
         # Fedora 29+ specific configuration file for Apache
         self.sysconfig_filep = "/etc/sysconfig/httpd"
         super().__init__(*args, **kwargs)